feat: Builder pattern 적용 + repository test 가 깨지지 않게 유지
