+++ src/main/java/nextstep/courses/domain/CourseRepository.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.course.Course;
+
 public interface CourseRepository {
     int save(Course course);
