+++ src/main/java/nextstep/courses/enums/CourseType.java
@@ -1,4 +1,4 @@
-package nextstep.qna.domain.enums;
+package nextstep.courses.enums;
 
 public enum CourseType {
     FREE, PAID
