--- src/main/java/nextstep/qna/domain/course/Course.java
@@ -1,16 +1,0 @@
-package nextstep.qna.domain.course;
-
-import nextstep.qna.domain.session.Session;
-
-import java.util.List;
-
-public class Course {
-
-    private Long id;
-
-    private String title;
-
-    private List<Session> sessions;
-
-
-}
