+++ src/test/java/nextstep/qna/domain/SessionTest.java
@@ -1,6 +1,6 @@
 package nextstep.qna.domain;
 
-import nextstep.qna.domain.session.Session;
+import nextstep.courses.domain.session.Session;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
